Content Helper: Move impreciseNumber() to shared functions
